backend/routers/auth.py
"""
CHE·NU™ V75 Backend - Auth Router

Authentication endpoints.

@version 75.0.0
"""

from fastapi import APIRouter, Depends, HTTPException, status, Request
from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
from sqlalchemy.ext.asyncio import AsyncSession
from datetime import datetime, timedelta
from typing import Optional
import jwt
import uuid
import logging

from config import settings, get_db
from schemas.auth import (
    LoginRequest,
    RegisterRequest,
    RefreshTokenRequest,
    AuthResponse,
    AuthTokens,
    User,
)
from schemas.base import BaseResponse, ErrorResponse, Meta

logger = logging.getLogger("chenu.auth")
router = APIRouter()
security = HTTPBearer(auto_error=False)


# ============================================================================
# JWT HELPERS
# ============================================================================

def create_access_token(user_id: str, roles: list = None) -> str:
    """Create JWT access token."""
    expires = datetime.utcnow() + timedelta(minutes=settings.JWT_ACCESS_TOKEN_EXPIRE_MINUTES)
    payload = {
        "sub": user_id,
        "roles": roles or ["user"],
        "exp": expires,
        "iat": datetime.utcnow(),
        "type": "access",
    }
    return jwt.encode(payload, settings.SECRET_KEY, algorithm=settings.JWT_ALGORITHM)


def create_refresh_token(user_id: str) -> str:
    """Create JWT refresh token."""
    expires = datetime.utcnow() + timedelta(days=settings.JWT_REFRESH_TOKEN_EXPIRE_DAYS)
    payload = {
        "sub": user_id,
        "exp": expires,
        "iat": datetime.utcnow(),
        "type": "refresh",
    }
    return jwt.encode(payload, settings.SECRET_KEY, algorithm=settings.JWT_ALGORITHM)


def decode_token(token: str) -> Optional[dict]:
    """Decode and validate JWT token."""
    try:
        payload = jwt.decode(
            token,
            settings.SECRET_KEY,
            algorithms=[settings.JWT_ALGORITHM]
        )
        return payload
    except jwt.ExpiredSignatureError:
        return None
    except jwt.InvalidTokenError:
        return None


async def get_current_user(
    credentials: HTTPAuthorizationCredentials = Depends(security),
    db: AsyncSession = Depends(get_db),
) -> Optional[dict]:
    """Get current user from JWT token."""
    if not credentials:
        return None
    
    payload = decode_token(credentials.credentials)
    if not payload or payload.get("type") != "access":
        return None
    
    # In production, fetch user from database
    # For now, return payload data
    return {
        "id": payload.get("sub"),
        "roles": payload.get("roles", ["user"]),
    }


def require_auth(user: dict = Depends(get_current_user)):
    """Dependency that requires authentication."""
    if not user:
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Not authenticated",
            headers={"WWW-Authenticate": "Bearer"},
        )
    return user


# ============================================================================
# ENDPOINTS
# ============================================================================

@router.post("/login", response_model=BaseResponse[AuthResponse])
async def login(
    request: LoginRequest,
    db: AsyncSession = Depends(get_db),
):
    """
    Authenticate user and return tokens.
    
    In production, this would:
    1. Validate email/password against database
    2. Check if user is active/verified
    3. Create session
    4. Return tokens
    """
    # TODO: Implement actual authentication
    # For now, mock response for development
    
    # Mock user (replace with database query)
    user_id = str(uuid.uuid4())
    user = User(
        id=user_id,
        email=request.email,
        name=request.email.split("@")[0],
        roles=["user"],
        permissions=["read", "write"],
        created_at=datetime.utcnow(),
    )
    
    tokens = AuthTokens(
        access_token=create_access_token(user_id, user.roles),
        refresh_token=create_refresh_token(user_id),
        expires_in=settings.JWT_ACCESS_TOKEN_EXPIRE_MINUTES * 60,
    )
    
    logger.info(f"User logged in: {request.email}")
    
    return BaseResponse(
        success=True,
        data=AuthResponse(user=user, tokens=tokens),
    )


@router.post("/register", response_model=BaseResponse[AuthResponse])
async def register(
    request: RegisterRequest,
    db: AsyncSession = Depends(get_db),
):
    """
    Register a new user.
    
    In production, this would:
    1. Check if email already exists
    2. Hash password
    3. Create user in database
    4. Send verification email
    5. Return tokens
    """
    # TODO: Implement actual registration
    
    user_id = str(uuid.uuid4())
    user = User(
        id=user_id,
        email=request.email,
        name=request.name,
        roles=["user"],
        permissions=["read", "write"],
        created_at=datetime.utcnow(),
    )
    
    tokens = AuthTokens(
        access_token=create_access_token(user_id, user.roles),
        refresh_token=create_refresh_token(user_id),
        expires_in=settings.JWT_ACCESS_TOKEN_EXPIRE_MINUTES * 60,
    )
    
    logger.info(f"User registered: {request.email}")
    
    return BaseResponse(
        success=True,
        data=AuthResponse(user=user, tokens=tokens),
    )


@router.post("/refresh", response_model=BaseResponse[AuthTokens])
async def refresh_token(
    request: RefreshTokenRequest,
    db: AsyncSession = Depends(get_db),
):
    """Refresh access token using refresh token."""
    payload = decode_token(request.refresh_token)
    
    if not payload or payload.get("type") != "refresh":
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Invalid refresh token",
        )
    
    user_id = payload.get("sub")
    
    # In production, validate user still exists and is active
    
    tokens = AuthTokens(
        access_token=create_access_token(user_id),
        refresh_token=create_refresh_token(user_id),
        expires_in=settings.JWT_ACCESS_TOKEN_EXPIRE_MINUTES * 60,
    )
    
    return BaseResponse(success=True, data=tokens)


@router.post("/logout")
async def logout(
    user: dict = Depends(require_auth),
    db: AsyncSession = Depends(get_db),
):
    """
    Logout user (invalidate session).
    
    In production, this would:
    1. Invalidate the current session
    2. Optionally blacklist the token
    """
    logger.info(f"User logged out: {user.get('id')}")
    
    return BaseResponse(
        success=True,
        data={"message": "Logged out successfully"},
    )


@router.get("/me", response_model=BaseResponse[User])
async def get_current_user_info(
    user: dict = Depends(require_auth),
    db: AsyncSession = Depends(get_db),
):
    """Get current user information."""
    # In production, fetch full user from database
    
    return BaseResponse(
        success=True,
        data=User(
            id=user.get("id"),
            email="user@example.com",  # Would come from DB
            name="Current User",
            roles=user.get("roles", ["user"]),
            permissions=["read", "write"],
            created_at=datetime.utcnow(),
        ),
    )


@router.get("/verify")
async def verify_token(
    credentials: HTTPAuthorizationCredentials = Depends(security),
):
    """Verify if token is valid."""
    if not credentials:
        return {"valid": False}
    
    payload = decode_token(credentials.credentials)
    if not payload:
        return {"valid": False}
    
    return {
        "valid": True,
        "user_id": payload.get("sub"),
        "expires_at": datetime.fromtimestamp(payload.get("exp")).isoformat(),
    }
