backend/requirements_unified.txt
# requirements.txt
# CHE·NU™ V75 - Backend Dependencies
#
# @version 75.0.0

# ============================================================================
# CORE FRAMEWORK
# ============================================================================
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ============================================================================
# DATABASE
# ============================================================================
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
aiosqlite==0.19.0

# ============================================================================
# REDIS
# ============================================================================
redis==5.0.1
aioredis==2.0.1

# ============================================================================
# HTTP CLIENT
# ============================================================================
httpx==0.26.0
aiohttp==3.9.1

# ============================================================================
# VALIDATION & SERIALIZATION
# ============================================================================
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# ============================================================================
# SECURITY
# ============================================================================
cryptography==41.0.7
itsdangerous==2.1.2

# ============================================================================
# UTILITIES
# ============================================================================
python-dotenv==1.0.0
orjson==3.9.10
ujson==5.9.0
python-dateutil==2.8.2

# ============================================================================
# LOGGING & MONITORING
# ============================================================================
structlog==24.1.0
sentry-sdk[fastapi]==1.39.1
prometheus-client==0.19.0

# ============================================================================
# FILE HANDLING
# ============================================================================
aiofiles==23.2.1
python-magic==0.4.27
Pillow==10.2.0

# ============================================================================
# WEBSOCKETS
# ============================================================================
websockets==12.0

# ============================================================================
# TESTING (Dev only)
# ============================================================================
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
factory-boy==3.3.0
faker==22.2.0

# ============================================================================
# CODE QUALITY (Dev only)
# ============================================================================
black==23.12.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
